+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Fixed bugs with sprite collisions
 - Added sprite dropshadow
 - Added gregium.setup for setting up files easily
+- Added gregium.gv for storing globals