+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-#### 0.1.9
+# Gregium Changelog
+
+## 0.1.9
+
 - Permanantly moved original gregium to gregium.old, creating new readme
 - Ported content of gregium.old to new gregium
 - Recursivley found packages and packagedata in setup